Spaceout listing to make entries clearer for user to view
THere is a global variable which can be used to control whether
the stocks summary displayed has space between individual companies
or not. By default it is set to True, so that there is a empty
line between successive companies.

If user wants to control the same at runtime, from within in the
running program they will have to use

"globals()['gbSpaceOutListing'] = False" or so ...