Trap and Handle statements which eval cant handle
As eval can only handle expressions, so any statements passed to it
will generate a syntax error. Trap these and see if exec can handle
them.

NOTE: However if the syntax error is due to other reasons like
error in the statement/expression, then it will lead to a double
exception for now.