Retain list_assets buysell da version (cur) also
revert from db to da, bcas I will be retaining this buy/sell da
version of the list logic.

Use the IBS for indexing into this bsda, and as part of same use
PRICE field name instead of VALUE.

The summary bsda logic's da is called da instead of assets, to be
in sync with convention used in other places.