Adb:Speedup importing of transactions by searching from end
As the db is maintained as a ascending time ordered structure, and
inturn as normally one will be inserting newer transactions, into
it, it will be faster to search for the right position in the db,
by searching from the end, rather than the begining. So the same
is done now.

This speeds up the importing of the trades into the system by
orders of magnitude. Because irrespective of the number of records,
chances are the right position will be found in o(1) time, mostly.